mod : refactorized mypage item list to uses common component
