+++ src/components/common/ItemList/ItemListBody/ItemButton.tsx
@@ -2,13 +2,13 @@
 import { css } from "@emotion/react";
 import { Button, Flex, Progress } from "antd";
 import { colors } from "../../../../styles/colors";
-import { ProductThumbnailInfo } from "../../../../models/product";
+import { CurrentUserBuy, CurrentUserSell, ProductThumbnailInfo } from "../../../../models/product";
 import { sm_lower_bound } from "../../../../styles/responsive";
 import { FireTwoTone } from "@ant-design/icons";
 import React from "react";
 
 interface ItemButtonProps {
-  product: ProductThumbnailInfo;
+  product: ProductThumbnailInfo | CurrentUserBuy | CurrentUserSell;
   showBidderCount?: boolean;
 }
