+++ src/models/product.ts
@@ -22,38 +22,40 @@ export interface ProductThumbnailInfo {
 
 export interface CurrentUserBuy {
   id: number;
-  product_name: string;
+  productName: string;
+  departmentName: string;
+  lowerBound: number;
+  currentHighestPrice: number;
+  upperBound: number;
+  imageUrl: string;
+  bidderCount?: number;
   user_id: number;
   status: string;
   user_highest_price: number;
-  lower_bound: number;
-  upper_bound: number;
   department_id: number;
-  departmentName: string;
   created_at: Date;
   updated_at: Date;
-  current_highest_price: number;
   image: {
     id: number;
     url: string;
   };
-  bidderCount: number;
 }
 export interface CurrentUserSell {
   id: number;
-  product_name: string;
+  productName: string;
+  departmentName: string;
+  lowerBound: number;
+  currentHighestPrice: number;
+  upperBound: number;
+  imageUrl: string;
+  bidderCount?: number;
   user_id: number;
   status: string;
-  lower_bound: number;
-  upper_bound: number;
   department_id: number;
-  departmentName: string;
   created_at: Date;
   updated_at: Date;
-  current_highest_price: number;
   image: {
     id: number;
     url: string;
   };
-  bidderCount: number;
 }
